+++ src/dummies/b-dummy/b-dummy.ts
@@ -37,7 +37,8 @@ import iData, {
 	field,
 	hook,
 	wait,
-	ModsNTable
+	ModsNTable,
+	VNode
 
 } from 'super/i-data/i-data';
 
@@ -45,6 +46,7 @@ import bBottomSlide from 'base/b-bottom-slide/b-bottom-slide';
 
 import daemons from 'dummies/b-dummy/daemons';
 import type { Directives, Modules, Engines } from 'dummies/b-dummy/interface';
+import type iBlock from 'super/i-block/i-block';
 
 const
 	inViewMutation = inViewFactory('mutation'),
@@ -64,8 +66,27 @@ interface bDummy extends Trait<typeof iLockPageScroll>, Trait<typeof iObserveDOM
 
 @derive(iLockPageScroll, iObserveDOM)
 class bDummy extends iData implements iLockPageScroll, iObserveDOM {
+
 	@field()
 	testField: unknown = undefined;
+
+	/**
+	 * Name of the test component
+	 */
+	@field()
+	testComponent?: string;
+
+	/**
+	 * Attributes for the test component
+	 */
+	@field()
+	testComponentAttrs: Dictionary = {};
+
+	/**
+	 * Slots for the test component
+	 */
+	@field()
+	testComponentSlots?: CanArray<VNode>;
 
 	get directives(): Directives {
 		return {
@@ -108,6 +129,10 @@ class bDummy extends iData implements iLockPageScroll, iObserveDOM {
 	}
 
 	static override readonly daemons: typeof daemons = daemons;
+
+	protected override readonly $refs!: iData['$refs'] & {
+		testComponent?: iBlock;
+	};
 
 	setStage(value: string): void {
 		this.stage = value;